Reorganize the code for dataset tests
