+++ mila_datamodules/vision/datasets/_utils.py
@@ -1,9 +1,16 @@
 from __future__ import annotations
 
+import inspect
 import itertools
 from pathlib import Path
-from typing import Iterable, Sequence
+from typing import Iterable, Protocol, Sequence, TypeVar
 
+import torchvision.datasets as tvd
+from torch.utils.data import Dataset
+from typing_extensions import TypeGuard
+
+DatasetType = TypeVar("DatasetType", bound=Dataset)
+VisionDatasetType = TypeVar("VisionDatasetType", bound=tvd.VisionDataset)
 ARCHIVE_FORMATS = ["*.zip", "*.gzip", "*.tar", "*.tar.gz", "*.hdf5"]
 METADATA_FORMATS = ["*.txt", "*.csv", "*.json"]
 
@@ -50,3 +57,14 @@ def glob_any(path: Path, patterns: Sequence[str]) -> Iterable[Path]:
 def rglob_any(path: Path, patterns: Sequence[str]) -> Iterable[Path]:
     """yields files matching any of the given patterns with `path.rglob(pattern)`."""
     return itertools.chain(*(path.rglob(pattern) for pattern in patterns))
+
+
+class DownloadableDataset(Protocol):
+    def __init__(self, root: str, *args, download: bool = False, **kwargs):
+        ...
+
+
+def is_downloadable(
+    dataset: type[VisionDatasetType],
+) -> TypeGuard[type[DownloadableDataset | VisionDatasetType]]:
+    return "download" in inspect.signature(dataset.__init__).parameters
